+++ ui/CHANGES.md
@@ -1,5 +1,6 @@
 #### Version: 1.1.5-SNAPSHOT (LibGDX 1.9.3)
 - **Changed**: It's now impossible to create `FileTypeFilter` `Rule` without providing at least one extension
+- **Changed**: `FileChooser` now can be closed by pressing enter when file name field has focus
 
 #### Version: 1.1.4 (LibGDX 1.9.3)
 - **Added**: `BusyBar` - used to indicate that background work is going on - see `TestBusyBar`